requirements.txt
fastapi
uvicorn
gunicorn
python-decouple
celery
redis
